// GC-Mobile-Web-Application/UnitedAdminPortal/src/app/models/professionalUser.ts
export class ProfessionalUser{
  userID: string;
  firstName: string;
  lastName: string;
  title: string;
  username: string;
  password: string;
  streetAddress: string;
  country: string;
  zipcode: string;
  contactNumber: string;
  email: string;
  profilePictureLink: string;
  calenderViewLink: string;
  secondaryProfilePictureLink: string;
  backStageCode: string;
  description : string;
}